pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.16"
asgiref = "^3.8.1"
attrs = "^24.2.0"
django-debug-toolbar = "^4.4.6"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
inflection = "^0.5.1"
jsonschema = "^4.23.0"
jsonschema-specifications = "^2024.10.1"
pillow = "^11.0.0"
pyyaml = "^6.0.2"
referencing = "^0.35.1"
rpds-py = "^0.21.0"
soupsieve = "^2.6"
sqlparse = "^0.5.2"
typing-extensions = "^4.12.2"
tzdata = "^2024.2"
uritemplate = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
